+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/model/YamlScript.java
@@ -1,22 +1,22 @@
 package com.walmartlabs.concord.project.yaml.model;
 
 import com.fasterxml.jackson.core.JsonLocation;
-import io.takari.bpm.model.ScriptTask;
+import io.takari.bpm.model.ScriptTask.Type;
 
 public class YamlScript extends YamlStep {
 
-    private final ScriptTask.Type type;
+    private final Type type;
     private final String language;
     private final String body;
 
-    public YamlScript(JsonLocation location, ScriptTask.Type type, String language, String body) {
+    public YamlScript(JsonLocation location, Type type, String language, String body) {
         super(location);
         this.type = type;
         this.language = language;
         this.body = body;
     }
 
-    public ScriptTask.Type getType() {
+    public Type getType() {
         return type;
     }