+++ project-model/src/test/resources/externalscript/test.js
@@ -1,0 +1,1 @@
+execution.setVariable("x", 12345);