+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ## Added
 
 - concord-console: support for "int" and "decimal" fields in the default form renderer;
-- added "project name" column to the process history.
+- added "project name" column to the process history;
+- concord-runner: support for external scripts.